Move the test framework out
