+++ test/apps/epoll/epoll_err.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
-#include "../network/test.h"
+#include "../test.h"
 #include <unistd.h>
 #include <sys/epoll.h>
