+++ test/apps/network/rtnl_err.c
@@ -3,7 +3,7 @@
 #include <netlink/netlink.h>
 #include <unistd.h>
 
-#include "test.h"
+#include "../test.h"
 
 static struct sockaddr_nl sk_addr = { .nl_family = AF_NETLINK };
