+++ test/apps/network/sockoption.c
@@ -7,7 +7,7 @@
 #include <netinet/tcp.h>
 #include <unistd.h>
 #include <arpa/inet.h>
-#include "test.h"
+#include "../test.h"
 
 int sk_unbound;
 int sk_listen;
