+++ test/apps/network/tcp_err.c
@@ -8,7 +8,7 @@
 #include <arpa/inet.h>
 #include <fcntl.h>
 
-#include "test.h"
+#include "../test.h"
 
 static struct sockaddr_in sk_addr;
