+++ test/apps/network/unix_err.c
@@ -11,7 +11,7 @@
 #include <unistd.h>
 #include <stddef.h>
 
-#include "test.h"
+#include "../test.h"
 
 FN_SETUP(general)
 {
