+++ test/apps/process/job_control.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
-#include "../network/test.h"
+#include "../test.h"
 
 #include <unistd.h>
 #include <pty.h>
